requirements.txt
discord.py
Pillow
pytesseract
python-dotenv
openai
google-generativeai